OE plot functions in separate file
